+++ src/onvif_media_snapshot.h
@@ -10,7 +10,7 @@ SHARD_EXPORT OnvifSnapshot * OnvifSnapshot__create(size_t size, char * buffer);
 SHARD_EXPORT void OnvifSnapshot__init(OnvifSnapshot * self, size_t size, char * buffer);
 SHARD_EXPORT void OnvifSnapshot__destroy(OnvifSnapshot * self);
 
-SHARD_EXPORT double OnvifSnapshot__get_size(OnvifSnapshot * self);
+SHARD_EXPORT int OnvifSnapshot__get_size(OnvifSnapshot * self);
 SHARD_EXPORT char * OnvifSnapshot__get_buffer(OnvifSnapshot * self);
 
 #endif
